+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Dependabot: Bump loader-utils from 1.4.1 to 1.4.2
+
 ## [0.1.2] - 2022-11-24
 
 ### Changed